Update roles controller - add try caches, move duplicated code in new function, add some new if clauses, improve security

+++ app/Http/Controllers/RoleController.php
@@ -8,6 +8,7 @@ use Spatie\Permission\Models\Role;
 use Spatie\Permission\Models\Permission;
 use DB;
 use Illuminate\Http\Response;
+use Illuminate\Support\Facades\Log;
 
 class RoleController extends Controller
 {
@@ -49,16 +50,20 @@ class RoleController extends Controller
      */
     public function store(Request $request)
     {
-        $this->validate($request, [
-            'name' => 'required|unique:roles,name',
-            'permission' => 'required',
-        ]);
-    
-        $role = Role::create(['name' => $request->input('name')]);
-        $role->syncPermissions($request->input('permission'));
-    
-        return redirect()->route('roles.index')
-                        ->with('success','Role created successfully');
+        if ($request->isMethod('post') && $request) {
+            $validated = $this->processValidate($request);
+            try {
+                $role = Role::create(['name' => $request->input('name')]);
+                $role->syncPermissions($request->input('permission'));
+            
+                return redirect()->route('roles.index')
+                                ->with('success','Role created successfully');
+            } catch (\Exception $e) {
+                //die($e->getMessage());
+                Log::error($e->getMessage());
+            }
+        }
+        return redirect()->route('roles.index');
     }
 
     /**
@@ -70,6 +75,10 @@ class RoleController extends Controller
     public function show($id)
     {
         $role = Role::find($id);
+        if (!$role) {
+            session()->flash('message', __('This role did not exists!'));
+            return redirect()->route('roles.create');
+        }
         $rolePermissions = Permission::join("role_has_permissions","role_has_permissions.permission_id","=","permissions.id")
             ->where("role_has_permissions.role_id", $id)
             ->get();
@@ -86,6 +95,10 @@ class RoleController extends Controller
     public function edit(int $id)
     {
         $role = Role::find($id);
+        if (!$role) {
+            session()->flash('message', __('This role did not exists!'));
+            return redirect()->route('roles.create');
+        }
         $permission = Permission::get();
         $rolePermissions = DB::table("role_has_permissions")->where("role_has_permissions.role_id", $id)
             ->pluck('role_has_permissions.permission_id','role_has_permissions.permission_id')
@@ -103,19 +116,22 @@ class RoleController extends Controller
      */
     public function update(Request $request, int $id)
     {
-        $this->validate($request, [
-            'name' => 'required',
-            'permission' => 'required',
-        ]);
-    
-        $role = Role::find($id);
-        $role->name = $request->input('name');
-        $role->save();
-    
-        $role->syncPermissions($request->input('permission'));
-    
-        return redirect()->route('roles.index')
-                        ->with('success','Role updated successfully');
+        if ($request->isMethod('patch') && $request) {
+            $validated = $this->processValidate($request);
+            try {
+                $role = Role::find($id);
+                $role->name = $request->input('name');
+                $role->save();
+                $role->syncPermissions($request->input('permission'));
+            
+                return redirect()->route('roles.index')
+                                ->with('success','Role updated successfully');
+            } catch (\Exception $e) {
+                Log::error($e->getMessage());
+                //die($e->getMessage());
+            }
+        }
+        return redirect()->route('roles.index');
     }
 
     /**
@@ -126,8 +142,25 @@ class RoleController extends Controller
      */
     public function destroy(int $id)
     {
-        DB::table("roles")->where('id',$id)->delete();
-        return redirect()->route('roles.index')
-                        ->with('success','Role deleted successfully');
+        if ($id) {
+            DB::table("roles")->where('id',$id)->delete();
+            return redirect()->route('roles.index')
+                            ->with('success','Role deleted successfully');
+        }
+        return redirect()->route('roles.index');
+    }
+    
+    /**
+     * process validate
+     *
+     * @param Request $request
+     * @return array
+     */
+    private function processValidate(Request $request): array
+    {
+        return $this->validate($request, [
+            'name' => 'required',
+            'permission' => 'required',
+        ]);
     }
 }
